+++ CHANGELOG.md
@@ -5,6 +5,16 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
+[Unreleased]
+
+### Added
+
+- New enum member `MessageEntityType.Unknown`
+
+### Fixed
+
+- Exception during deserialization of unknown message entity type  
+
 ## [14.4.0] - 2018-05-17
 
 ### Changed